Fix handler_decrease converting the amount to int instead of a float

+++ illusion.py
@@ -194,7 +194,7 @@ class DB_Commands:
 
         sku = illusion_helpers.clean_sku(sku)
 
-        if amount != None and int(amount) <= 0:
+        if amount != None and float(amount) <= 0:
             return f"Quantity must be greater than 0"
         elif inventory.validate_sku(sku):
             result = inventory.decrease_item(sku, amount)
